Changed 2fa factors id types

+++ Example/TokenDSDK/ApiExampleViewController.swift
@@ -708,9 +708,12 @@ class ApiExampleViewController: UIViewController, RequestSignKeyDataProviderProt
                     
                     let totpFactors = factors.getTOTPFactors()
                     if let totpFactor = totpFactors.first {
+                        guard let id = Int(totpFactor.id) else {
+                            return
+                        }
                         self.tokenDApi.tfaApi.deleteFactor(
                             walletId: walletId,
-                            factorId: totpFactor.id,
+                            factorId: id,
                             completion: { (deleteResult) in
                                 switch deleteResult {
                                     
@@ -997,8 +1000,10 @@ class ApiExampleViewController: UIViewController, RequestSignKeyDataProviderProt
             style: .default,
             handler: { _ in
                 UIPasteboard.general.string = response.attributes.secret
-                
-                self.updateTOTPFactor(walletId: walletId, factorId: response.id, priority: priority)
+                guard let id = Int(response.id) else {
+                    return
+                }
+                self.updateTOTPFactor(walletId: walletId, factorId: id, priority: priority)
         }))
         
         if let url = URL(string: response.attributes.seed),
@@ -1009,8 +1014,10 @@ class ApiExampleViewController: UIViewController, RequestSignKeyDataProviderProt
                 handler: { _ in
                     UIPasteboard.general.string = response.attributes.secret
                     UIApplication.shared.open(url, options: [:], completionHandler: nil)
-                    
-                    self.updateTOTPFactor(walletId: walletId, factorId: response.id, priority: priority)
+                    guard let id = Int(response.id) else {
+                        return
+                    }
+                    self.updateTOTPFactor(walletId: walletId, factorId: id, priority: priority)
             }))
         }
